+++ types/src/weekUtils.d.ts
@@ -1,0 +1,3 @@
+export declare function getWeekDaysNames(): string[];
+export declare function getShortWeekDaysNames(): string[];
+export declare function getNarrowWeekDaysNames(): string[];