+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 1.2.0
+
+- Feature: add "monthUtils" with functions to retrieve the months names in the user's locale
+- Feature: add "weekUtils" with functions to retrieve the week days' names in the user's locale
+- Feature: add temparature helper to obtain temprature units and convert them
+- Feature: add temperature formatter
+
 # 1.1.1
 
 - Fix: "numberFormatter" missing from library